+++ CHANGES.md
@@ -21,6 +21,8 @@ CHANGES IN CUPS V2.2.5
   printers (rdar://33250434)
 - Fixed the `cups.strings` file that is used on macOS (rdar://33287650)
 - CUPS now sends the `Date` HTTP header in IPP requests (rdar://33302034)
+- The `cupsGetNamedDest` function did not use the local default printer
+  (rdar://33228500)
 
 
 CHANGES IN CUPS V2.2.4